1.0.0-beta-sun-build20 Version:
* Core package update
* Addition of request extensions class
* Addition of response class definition
* Remove of ISingleResponse for put and delete actions

Quick start: https://www.codeproject.com/Tips/1229909/Scaffolding-ASP-NET-Core-with-CatFactory

+++ CatFactory.AspNetCore/Definitions/Extensions/PagedResponseInterfaceBuilder.cs
@@ -28,7 +28,7 @@ namespace CatFactory.AspNetCore.Definitions.Extensions
                 Properties =
                 {
                     new PropertyDefinition("int", "ItemsCount"),
-                    new PropertyDefinition("int", "PageCount")
+                    new PropertyDefinition("double", "PageCount")
                 }
             };
     }
